travel-planner-agent/requirements.txt
langchain>=0.0.26
openai>=0.27.0
python-dotenv>=0.19.0
anthropic>=0.18.0
tushare>=1.2.77

# LangChain 核心依赖
# langchain
langchain-openai>=0.0.1
langchain-community>=0.0.1

# LLM 提供商
# openai - OpenAI API
# anthropic - Anthropic Claude API

# 数据处理
# pandas - 数据分析

# 股市数据
# tushare - 中国股市数据

# 异步支持
# asyncio - 异步 I/O

# 日志和监控
# tenacity - 重试机制
